Moved config(Info) files to respective HCDs && changed dependencies to avoid circular dependencies
