+++ wfos_ics/build.sbt
@@ -12,24 +12,22 @@ lazy val `wfos-ics-root` = project
 
 // assembly module
 lazy val `wfos-bgrxassembly` = project
+  .dependsOn(
+    `wfos-lgripHcd`,
+    `wfos-rgripHcd`
+  )
   .settings(
     libraryDependencies ++= Dependencies.bgrxassembly
   )
 
 // hcd module
 lazy val `wfos-lgripHcd` = project
-  .dependsOn(
-    `wfos-bgrxassembly`
-  )
   .settings(
     libraryDependencies ++= Dependencies.lgripHcd
   )
 
   // hcd module
 lazy val `wfos-rgripHcd` = project
-  .dependsOn(
-    `wfos-bgrxassembly`
-  )
   .settings(
     libraryDependencies ++= Dependencies.rgripHcd
   )
