+++ wfos_ics/wfos-bgrxassembly/src/main/scala/wfos/bgrxassembly/components/LgripHcd.scala
@@ -3,7 +3,7 @@ package wfos.bgrxassembly.components
 import csw.params.commands.{CommandIssue, Setup}
 import csw.params.commands.CommandIssue.{ParameterValueOutOfRangeIssue, MissingKeyIssue}
 import csw.params.core.generics.{Parameter}
-import wfos.bgrxassembly.config.LgripInfo
+import wfos.lgriphcd.LgripInfo
 
 class LgripHcd extends Hcd {
   override def validateParameters(setup: Setup): Either[CommandIssue, Parameter[Int]] = {
